assign ambassador automatic after three days
