+++ Limbs.Web.Logic/Helpers/SiteHelper.cs
@@ -4,7 +4,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Limbs.Web.Helpers
+namespace Limbs.Web.Logic.Helpers
 {
     public static class SiteHelper
     {
