+++ Limbs.Web.Logic/packages.config
@@ -21,7 +21,12 @@
   <package id="System.ComponentModel.EventBasedAsync" version="4.0.11" targetFramework="net461" />
   <package id="System.Dynamic.Runtime" version="4.0.0" targetFramework="net461" />
   <package id="System.Linq.Queryable" version="4.0.0" targetFramework="net461" />
+  <package id="System.Net.Http" version="4.3.4" targetFramework="net461" />
   <package id="System.Net.Requests" version="4.0.11" targetFramework="net461" />
+  <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net461" />
+  <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net461" />
+  <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net461" />
+  <package id="System.Security.Cryptography.X509Certificates" version="4.3.0" targetFramework="net461" />
   <package id="System.Spatial" version="5.8.2" targetFramework="net461" />
   <package id="WindowsAzure.Storage" version="8.7.0" targetFramework="net461" />
 </packages>
