+++ Limbs.Worker/packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="EntityFramework" version="6.2.0" targetFramework="net461" />
+  <package id="LightInject" version="5.1.2" targetFramework="net461" />
   <package id="Microsoft.Azure.KeyVault.Core" version="2.0.4" targetFramework="net461" />
   <package id="Microsoft.Azure.WebJobs" version="2.3.0" targetFramework="net461" />
   <package id="Microsoft.Azure.WebJobs.Core" version="2.3.0" targetFramework="net461" />
